Fully qualify method names automatically
inside lambda on Immediate Pad. For example, if the current instance of the class
has an instance metthod, and an user inputs `(x => InstanceMethod(x))`, it will be
evaluated as `(x => this.InstanceMethod(x))`. Similarly, a lambda expression
`(x => StaticMethod(x))` in static context of `ClassA` will become
`(x => ClassA.StaticMethod(x))`